# Improve image crop performance