@pierrejoye 2022-01-23
thanks for this nice PR :)

Any objection to merge?